Fix G_RunFrame mixing frametime and serverTime
The old code added frametime to game.serverTime-relative variables.
Those two timers are not synchronized, and there were skips when a frame
took more than 16ms (for example, when the game is in background). This
is especially noticeable if you call vid_restart while waiting for
players or after firing a rocket and calling a timeout.