Don't require a manual vid_restart if rserr_restart_required happens
In the previous versions, vid_fullscreen was reset when
rserr_restart_required was returned from VID_ChangeMode, thus making the
text about vid_restart wrong and misleading.

This change also makes Alt+Enter work in all cases for the players, no
matter whether a restart is required or not, and automatically fixes the
screen resolution if it's unsupported by the fullscreen mode.