Fix platform movement for real
It was broken since the very beginning of AAS nav introduction.
Losing a nav target standing on a platform is a primary issue.
It has been solved similar to the jumppad movement handling.
A list of areas usable to exit a platform is built and used
when the platform reaches TOP state.