Rework caching of area and cluster memory blocks
Allocation of a freelist pool with predefined size
has proven to be a bad solution, since a block size
exceeds the predefined size on several maps (like wctf2).
This leads to falling back to raw G_Malloc() calls
and, consequently, to poor performance and crashes
due to memory fragmentation.

A better solution is implemented. A separate cache bin is
used for each possible size of a block (there are only few
possible block sizes, depending of a map). Bins are arranged
in a lookup table or in a linked list, depending of
an actual block size.

A bin allocates a freelist pool based on a continuous memory
region lazily, since for some block sizes only a single
block is needed. Otherwise raw G_Malloc() call is used, and
its result is cached for further using. Once extra blocks
are requested for allocation, a freelist pool is created
and further allocations are redirected to it.